src/test/kotlin/me/sivieri/aoc2020/day20/TilesTest.kt
package me.sivieri.aoc2020.day20

import org.junit.Assert
import org.junit.FixMethodOrder
import org.junit.Test
import org.junit.runners.MethodSorters

@FixMethodOrder(MethodSorters.NAME_ASCENDING)
class TilesTest {

    private val input = """
            Tile 2311:
            ..##.#..#.
            ##..#.....
            #...##..#.
            ####.#...#
            ##.##.###.
            ##...#.###
            .#.#.#..##
            ..#....#..
            ###...#.#.
            ..###..###

            Tile 1951:
            #.##...##.
            #.####...#
            .....#..##
            #...######
            .##.#....#
            .###.#####
            ###.##.##.
            .###....#.
            ..#.#..#.#
            #...##.#..

            Tile 1171:
            ####...##.
            #..##.#..#
            ##.#..#.#.
            .###.####.
            ..###.####
            .##....##.
            .#...####.
            #.##.####.
            ####..#...
            .....##...

            Tile 1427:
            ###.##.#..
            .#..#.##..
            .#.##.#..#
            #.#.#.##.#
            ....#...##
            ...##..##.
            ...#.#####
            .#.####.#.
            ..#..###.#
            ..##.#..#.

            Tile 1489:
            ##.#.#....
            ..##...#..
            .##..##...
            ..#...#...
            #####...#.
            #..#.#.#.#
            ...#.#.#..
            ##.#...##.
            ..##.##.##
            ###.##.#..

            Tile 2473:
            #....####.
            #..#.##...
            #.##..#...
            ######.#.#
            .#...#.#.#
            .#########
            .###.#..#.
            ########.#
            ##...##.#.
            ..###.#.#.

            Tile 2971:
            ..#.#....#
            #...###...
            #.#.###...
            ##.##..#..
            .#####..##
            .#..####.#
            #..#.#..#.
            ..####.###
            ..#.#.###.
            ...#.#.#.#

            Tile 2729:
            ...#.#.#.#
            ####.#....
            ..#.#.....
            ....#..#.#
            .##..##.#.
            .#.####...
            ####.#.#..
            ##.####...
            ##..#.##..
            #.##...##.

            Tile 3079:
            #.#.#####.
            .#..######
            ..#.......
            ######....
            ####.#..#.
            .#...#.##.
            #.#####.##
            ..#.###...
            ..#.......
            ..#.###...
        """.trimIndent()
        .split("\n")

    private val image = """
        .#.#..#.##...#.##..#####
        ###....#.#....#..#......
        ##.##.###.#.#..######...
        ###.#####...#.#####.#..#
        ##.#....#.##.####...#.##
        ...########.#....#####.#
        ....#..#...##..#.#.###..
        .####...#..#.....#......
        #..#.##..#..###.#.##....
        #.####..#.####.#.#.###..
        ###.#.#...#.######.#..##
        #.####....##..########.#
        ##..##.#...#...#.#.#.#..
        ...#..#..#.#.##..###.###
        .#.#....#.##.#...###.##.
        ###.#...#..#.##.######..
        .#.#.###.##.##.#..#.##..
        .####.###.#...###.#..#.#
        ..#.#..#..#.#.#.####.###
        #..####...#.#.#.###.###.
        #####..#####...###....##
        #.##..#..#...#..####...#
        .#.###..##..##..####.##.
        ...###...##...#...#..###
    """.trimIndent()

    @Test
    fun `01 find angles`() {
        val tiles = Tiles(input)
        val res = tiles.getAnglesIdProduct()
        Assert.assertEquals(20899048083289L, res)
    }

    @Test
    fun `02 solve puzzle`() {
        val tiles = Tiles(input)
        val res = tiles.solvePuzzle()
        val exp = listOf(
            listOf(2971, 1489, 1171),
            listOf(2729, 1427, 2473),
            listOf(1951, 2311, 3079)
        )
        Assert.assertEquals(exp, res)
    }

    @Test
    fun `03 compose image`() {
        val tiles = Tiles(input)
        tiles.solvePuzzle()
        val img = tiles
            .prepareImage()
        val res = Matrix
            .mirrorHorizontal(
                img
                    .toList()
                    .subList(0, img.size - 1)
                    .toTypedArray()
            )
            .joinToString("\n") {
                it.joinToString("")
            }
        Assert.assertEquals(image, res)
    }

    @Test
    fun `04 find rough waters`() {
        val tiles = Tiles(input)
        tiles.solvePuzzle()
        val roughWaters = tiles.findRoughWaters()
        Assert.assertEquals(273, roughWaters)
    }

}